requirements.txt
streamlit
pandas
numpy
scikit-learn
scipy
seaborn
matplotlib
pandas_profiling
streamlit_pandas_profilin